Changed the "~>" form and added the "set!/chor" form.
Changed the "~>" form so that it no longer supports the syntax or functionality
of "(~> (at l1 x) (at l2 y))", that is it no longer directly supports mutating
a variables (such as y above). The "(~> (at l1 x) (at l2))" form was kept (which
evaluates to a choreographic value) but its syntax was changed to
"(~> (at l1 x) l2)".

To make up for the removal of the first "~>" form, the "set!/chor" was added.