Fixed bug with context based expansion. Added syntax classes.
Inside of "at" forms (specifically those which evaluate a local expression) the
"in-global-expr" syntax parameter was not being correctly set to "#f", allowing
Choret forms to be incorrectly used inside of local expressions. As part of
fixing this bug, two syntax classes, "at-id" and "at-expr" where added, and two
tests were added.