Added two examples and have them used as tests.
Added two new examples "book-seller.rkt", which is an example based off of the
Pirouette paper, and "factorial.rkt", which is based off one of the old examples
from the simple-choreographies implementation. Also added the ability to use
these examples (and other examples that are added in the future) as part of the
tests for Choret.