Fixed bug with merging.
There was a bug with merging where locations which were not part of a selection
would project "(void)" for that selection instead of the expression inside the
selection. For example, the the expansion of "(sel~> A [B 'b (at C 0)])" during
the projection of location "C" would incorrectly be "(void)" instead of
"(at C 0)". This bug was noticed when creating tests for nested merges.