Changed the syntax and semantics of the "sel~>" selection form.
Changed the syntax from:
  (sel~> SENDER [RECIEVER LABEL GEXPR] ...)
to:
  (sel~> SENDER ([RECIEVER LABEL] ...) GEXPR)

The semantics also changed slightly. Before the syntax with multiple
receiver-label pairs were roughly equivalent to multiple separate "sel~>" forms,
one for each receiver-label-gexpr tuple, which was not very useful since:
1) It could easily be replicated by just using multiple "sel~>" forms.
2) Each gexpr was bound to a single receiver-label pair, which did allow nested
   selections to be encoded, instead a "sel~>" form would have to be manually
   nested inside another "sel~>" form.

With the new syntax multiple receiver-label pairs are the same as having nested
selections, and in the single gexpr the knowledge of choice for all the listed
receiver locations is available.